requirements.txt
pyyaml==6.0.2
networkx==3.4.2
matplotlib==3.10.3